# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * wearurway API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Product {
  id: string;
  name: string;
  available: boolean;
  comingSoon: boolean;
}

export interface UpdateProductBody {
  available?: boolean;
  comingSoon?: boolean;
}

export interface Fit {
  id: string;
  name: string;
  available: boolean;
  comingSoon: boolean;
  productId: string;
}

export interface UpdateFitBody {
  available?: boolean;
  comingSoon?: boolean;
}

export interface Color {
  id: string;
  name: string;
  hex: string;
  fitId: string;
}

export interface AddColorBody {
  name: string;
  hex: string;
}

export interface Size {
  id: string;
  name: string;
  realWidth: number;
  realHeight: number;
  image?: string;
  fitId: string;
}

export interface AddSizeBody {
  name: string;
  realWidth: number;
  realHeight: number;
  image?: string;
}

export interface UpdateSizeBody {
  name?: string;
  realWidth?: number;
  realHeight?: number;
  image?: string;
}

export interface AdminLoginBody {
  password: string;
}

export interface AdminLoginResponse {
  success: boolean;
  message: string;
}

export interface AdminMeResponse {
  authenticated: boolean;
}

export interface ErrorResponse {
  error: string;
}