# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * wearurway API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Get all products
 */
export const GetProductsResponseItem = zod.object({
  id: zod.string(),
  name: zod.string(),
  available: zod.boolean(),
  comingSoon: zod.boolean(),
});
export const GetProductsResponse = zod.array(GetProductsResponseItem);

/**
 * @summary Update a product (admin)
 */
export const UpdateProductParams = zod.object({
  id: zod.coerce.string(),
});

export const UpdateProductBody = zod.object({
  available: zod.boolean().optional(),
  comingSoon: zod.boolean().optional(),
});

export const UpdateProductResponse = zod.object({
  id: zod.string(),
  name: zod.string(),
  available: zod.boolean(),
  comingSoon: zod.boolean(),
});

/**
 * @summary Get all fits
 */
export const GetFitsResponseItem = zod.object({
  id: zod.string(),
  name: zod.string(),
  available: zod.boolean(),
  comingSoon: zod.boolean(),
  productId: zod.string(),
});
export const GetFitsResponse = zod.array(GetFitsResponseItem);

/**
 * @summary Update a fit (admin)
 */
export const UpdateFitParams = zod.object({
  id: zod.coerce.string(),
});

export const UpdateFitBody = zod.object({
  available: zod.boolean().optional(),
  comingSoon: zod.boolean().optional(),
});

export const UpdateFitResponse = zod.object({
  id: zod.string(),
  name: zod.string(),
  available: zod.boolean(),
  comingSoon: zod.boolean(),
  productId: zod.string(),
});

/**
 * @summary Get colors for a fit
 */
export const GetColorsParams = zod.object({
  fitId: zod.coerce.string(),
});

export const GetColorsResponseItem = zod.object({
  id: zod.string(),
  name: zod.string(),
  hex: zod.string(),
  fitId: zod.string(),
});
export const GetColorsResponse = zod.array(GetColorsResponseItem);

/**
 * @summary Add a color to a fit (admin)
 */
export const AddColorParams = zod.object({
  fitId: zod.coerce.string(),
});

export const AddColorBody = zod.object({
  name: zod.string(),
  hex: zod.string(),
});

/**
 * @summary Remove a color from a fit (admin)
 */
export const DeleteColorParams = zod.object({
  fitId: zod.coerce.string(),
  colorId: zod.coerce.string(),
});

/**
 * @summary Get sizes for a fit
 */
export const GetSizesParams = zod.object({
  fitId: zod.coerce.string(),
});

export const GetSizesResponseItem = zod.object({
  id: zod.string(),
  name: zod.string(),
  realWidth: zod.number(),
  realHeight: zod.number(),
  image: zod.string().optional(),
  fitId: zod.string(),
});
export const GetSizesResponse = zod.array(GetSizesResponseItem);

/**
 * @summary Add a size to a fit (admin)
 */
export const AddSizeParams = zod.object({
  fitId: zod.coerce.string(),
});

export const AddSizeBody = zod.object({
  name: zod.string(),
  realWidth: zod.number(),
  realHeight: zod.number(),
  image: zod.string().optional(),
});

/**
 * @summary Update a size (admin)
 */
export const UpdateSizeParams = zod.object({
  fitId: zod.coerce.string(),
  sizeId: zod.coerce.string(),
});

export const UpdateSizeBody = zod.object({
  name: zod.string().optional(),
  realWidth: zod.number().optional(),
  realHeight: zod.number().optional(),
  image: zod.string().optional(),
});

export const UpdateSizeResponse = zod.object({
  id: zod.string(),
  name: zod.string(),
  realWidth: zod.number(),
  realHeight: zod.number(),
  image: zod.string().optional(),
  fitId: zod.string(),
});

/**
 * @summary Delete a size (admin)
 */
export const DeleteSizeParams = zod.object({
  fitId: zod.coerce.string(),
  sizeId: zod.coerce.string(),
});

/**
 * @summary Admin login
 */
export const AdminLoginBody = zod.object({
  password: zod.string(),
});

export const AdminLoginResponse = zod.object({
  success: zod.boolean(),
  message: zod.string(),
});

/**
 * @summary Check admin session
 */
export const GetAdminMeResponse = zod.object({
  authenticated: zod.boolean(),
});