# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * wearurway API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: string;
  name: string;
  available: boolean;
  comingSoon: boolean;
}