# lib/api-zod/src/generated/types/updateFitBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * wearurway API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateFitBody {
  available?: boolean;
  comingSoon?: boolean;
}