# lib/api-zod/src/generated/types/updateSizeBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * wearurway API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateSizeBody {
  name?: string;
  realWidth?: number;
  realHeight?: number;
  image?: string;
}